wiki search base concept working
The DeepThoughtHelper.py now succeds in it's goal.

Search through a wiki page for the number 42 and the sentence that contains it.

Iterate to the next wiki page alphabetically.

The next step is to make a batch file that will run this one iteration after another on and on until we go through every wiki page.